authorized users to delete only

+++ events/views.py
@@ -4,7 +4,7 @@ from calendar import HTMLCalendar
 from datetime import datetime
 from . models import Event, Venue
 from .forms import VenueForm, EventForm, EventFormAdmin
-
+from django.contrib import messages
 from django.http import HttpResponseRedirect
 from django.http import HttpResponse
 import csv
@@ -96,8 +96,14 @@ def delete_venue(request, venue_id):
 # Delete an event
 def delete_event(request, event_id):
     event  = Event.objects.get(pk=event_id)
-    event.delete()
-    return redirect('list-events')
+    if request.user == event.manager:
+        event.delete()
+        messages.success(request, 'Event Deleted!')
+
+        return redirect('list-events')
+    else:
+        messages.success(request, "You aren't Authorized To Delete This Event!")
+        return redirect('list-events')
 
 
 def update_event(request, event_id):
